bs_project/requirements.txt
Flask==3.0.2
PyJWT==2.8.0
psycopg2==2.9.9
python-dotenv==1.0.1
